+++ R/convertUSCensusStates.R
@@ -1,5 +1,6 @@
 #' @keywords datagen
 #' @importFrom rlang .data
+#' @importFrom cleangeo clgeo_IsValid
 #' @export
 #' 
 #' @title Convert US Census State Shapefile
@@ -148,7 +149,7 @@ convertUSCensusStates <- function(
       AFFGEOID = .data$AFFGEOID
     )
   
-  # ----- Organize polygons ----------------------------------------------------
+  # ----- Clean SPDF -----------------------------------------------------------
   
   # Group polygons with the same identifier (stateFIPS)
   SPDF <- organizePolygons(
@@ -156,6 +157,11 @@ convertUSCensusStates <- function(
     uniqueID = 'stateFIPS', 
     sumColumns = c('areaLand', 'areaWater')
   )
+  
+  # Clean topology errors
+  if ( !cleangeo::clgeo_IsValid(SPDF) ) {
+    SPDF <- cleangeo::clgeo_Clean(SPDF)
+  }
   
   # ----- Name and save the data -----------------------------------------------
   
@@ -173,28 +179,40 @@ convertUSCensusStates <- function(
     message("Simplifying to 5%...\n")
     SPDF_05 <- rmapshaper::ms_simplify(SPDF, 0.05)
     SPDF_05@data$rmapshaperid <- NULL # Remove automatically generated "rmapshaperid" column
+    # Clean topology errors
+    if ( !cleangeo::clgeo_IsValid(SPDF_05) ) {
+      SPDF_05 <- cleangeo::clgeo_Clean(SPDF_05)
+    }
     datasetName_05 <- paste0(datasetName, "_05")
     message("Saving 5% version...\n")
     assign(datasetName_05, SPDF_05)
-    save(list = datasetName_05, file = paste0(dataDir,"/",datasetName_05, '.rda'))
+    save(list = datasetName_05, file = paste0(dataDir,"/", datasetName_05, '.rda'))
     rm(list = c("SPDF_05",datasetName_05))
     
     message("Simplifying to 2%...\n")
     SPDF_02 <- rmapshaper::ms_simplify(SPDF, 0.02)
     SPDF_02@data$rmapshaperid <- NULL # Remove automatically generated "rmapshaperid" column
+    # Clean topology errors
+    if ( !cleangeo::clgeo_IsValid(SPDF_02) ) {
+      SPDF_02 <- cleangeo::clgeo_Clean(SPDF_02)
+    }
     datasetName_02 <- paste0(datasetName, "_02")
     message("Saving 2% version...\n")
     assign(datasetName_02, SPDF_02)
-    save(list = datasetName_02, file = paste0(dataDir,"/",datasetName_02, '.rda'))
+    save(list = datasetName_02, file = paste0(dataDir,"/", datasetName_02, '.rda'))
     rm(list = c("SPDF_02",datasetName_02))
     
     message("Simplifying to 1%...\n")
     SPDF_01 <- rmapshaper::ms_simplify(SPDF, 0.01)
     SPDF_01@data$rmapshaperid <- NULL # Remove automatically generated "rmapshaperid" column
+    # Clean topology errors
+    if ( !cleangeo::clgeo_IsValid(SPDF_01) ) {
+      SPDF_01 <- cleangeo::clgeo_Clean(SPDF_01)
+    }
     datasetName_01 <- paste0(datasetName, "_01")
     message("Saving 1% version...\n")
     assign(datasetName_01, SPDF_01)
-    save(list = datasetName_01, file = paste0(dataDir,"/",datasetName_01, '.rda'))
+    save(list = datasetName_01, file = paste0(dataDir,"/", datasetName_01, '.rda'))
     rm(list = c("SPDF_01",datasetName_01))
   }